created authorization with tokens

+++ controllers/blog.js
@@ -2,6 +2,16 @@ const blogsRouter = require('express').Router();
 const Blogs = require('../models/blogs');
 const User = require('../models/user');
 
+const jwt = require('jsonwebtoken');
+
+const getTokenFrom = (request) => {
+  const authorization = request.get('authorization');
+  if (authorization && authorization.startsWith('Bearer ')) {
+    return authorization.replace('Bearer ', '');
+  }
+  return null;
+};
+
 blogsRouter.get('/', async (request, response) => {
   const blog = await Blogs.find({}).populate('user', { username: 1, name: 1 });
   response.json(blog);
@@ -27,7 +37,12 @@ blogsRouter.post('/', async (request, response) => {
     return response.status(400).json({ error: 'Missing required fields' });
   }
 
-  const user = await User.findById(body.userId);
+  const decodedToken = jwt.verify(getTokenFrom(request), process.env.SECRET);
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' });
+  }
+
+  const user = await User.findById(decodedToken.id);
 
   const blog = new Blogs({
     title: body.title,
@@ -45,9 +60,29 @@ blogsRouter.post('/', async (request, response) => {
 });
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blogs.findByIdAndRemove(request.params.id);
-  response.status(204).end();
+  try {
+    const decodedToken = jwt.verify(getTokenFrom(request), process.env.SECRET);
+    if (!decodedToken.id) {
+      return response.status(401).json({ error: 'Token invalid' });
+    }
+    
+    const blog = await Blogs.findById(request.params.id);
+    
+    if (!blog) {
+      return response.status(404).json({ error: 'Blog not found' });
+    }
+
+    if (blog.user.toString() !== decodedToken.id) {
+      return response.status(403).json({ error: 'Unauthorized to delete this blog' });
+    }
+
+    await Blogs.findByIdAndRemove(request.params.id);
+    response.status(204).end();
+  } catch (error) {
+    response.status(500).json({ error: 'Server error' });
+  }
 });
+
 
 blogsRouter.put('/:id', async (request, response, next) => {
   const body = request.body;
